# Optimize percent-decode