Stop re-parsing the percent-encoded chunk with `re-seq`. Also remove intermediate sequences.