@weavejester 2022-02-02
Thanks for the PR. Why did you inline `parse-bytes`? Why not put the loop in `parse-bytes`? Do you have benchmarks for performance before and after?