@nilern 2022-02-02
I thought it better to inline a function with such an awkward precondition ("must match this regex..."). Although I suppose it already had that precondition, so it was a bit orthogonal to the optimization. In any case, I should have mentioned this.

A quick bench based on some test example was about 50% faster. Basically I just checked that it was actually faster, I could benchmark more if you want.