@weavejester 2022-02-03
Rather than inlining `parse-bytes`, can you instead put the new code in that function instead? Could you also make the internal variables (like `ci` and `bi`) a little more descriptive?

If you test the before and after function with Criterium, can you also add the output of that to this PR as a comment so we have a reference that people can refer back to in future?