@nilern 2022-02-05
The results from my sanity benchmark:

# Before

```clojure
user=> (crit/bench (codec/percent-decode "foo%FE%FF%00%2Fbar" "UTF-16"))
Evaluation count : 618960 in 60 samples of 10316 calls.
             Execution time mean : 97,466064 µs
    Execution time std-deviation : 1,427946 µs
   Execution time lower quantile : 96,070529 µs ( 2,5%)
   Execution time upper quantile : 99,159458 µs (97,5%)
                   Overhead used : 13,943488 ns

Found 1 outliers in 60 samples (1,6667 %)
	low-severe	 1 (1,6667 %)
 Variance from outliers : 1,6389 % Variance is slightly inflated by outliers
```
# After

```clojure
user=> (crit/bench (codec/percent-decode "foo%FE%FF%00%2Fbar" "UTF-16"))
Evaluation count : 1216500 in 60 samples of 20275 calls.
             Execution time mean : 49,266600 µs
    Execution time std-deviation : 2,942700 µs
   Execution time lower quantile : 47,739405 µs ( 2,5%)
   Execution time upper quantile : 58,222833 µs (97,5%)
                   Overhead used : 13,601526 ns

Found 8 outliers in 60 samples (13,3333 %)
	low-severe	 2 (3,3333 %)
	low-mild	 6 (10,0000 %)
 Variance from outliers : 45,0770 % Variance is moderately inflated by outliers
```