@weavejester 2022-02-09
Thanks! Could you squash your two commits down? Then I'll merge it in and cut a release.